Added SDIO control APB register:
this was added as the LInux GPIO driver was resetting GPIOs during boot and disconnecting the SD card resulting in rootfs failing to mount. Using the register no drivers should cause it to reset. The regitser can be set by writing a value of 1 and unset by writing a value of 0 or causing a reset. It can be connected to an APB bus and can accept 32 bit transactions.

Instansiated coreAPB3, instansiated and connected the register to slot 15 (address 0x4f00_0000).

Libero is generating a configuration file for the register in the top level script folder instead of in the project when generating the core. I've added this file to the git ignore to stop it being included in the commit. Investigating why its doing this.