Clocks: Update CCC to use 50MHz ref clk
The CCC was updated to use the on board 50MHz reference clock instead of
the 160MHz osc. Updated the Clocks and Resets smartdesign to add the
additional input port for this clock and then also added it to the top
level smartdesign. Added the constraint for the 50MHz clock to the
ICICLE constraints file