argument design: add MICRON_QSPI argument design
Added a "Micron_QSPI" argument to assocaite a different constraint for
use with Micron QSPI devices.

The main Tcl script is updated to add the argument and assocaite
constraints. Added a new constraint file which is associated if the new
argument is passed. Updated the gitignore to ignore the new project
folder that is created.