CI: Split AXI4_STREAM_DEMO into two parts
The testing for the AXI4_STREAM_DEMO is split into two parts so the hardware design is generated and a job file exported and then separately the BFM design is generated.

The BFM design no longer runs EXPORT_FPE or HSS_UPDATE as arguments as this design is intended to demo BFM simulation and not be run on hardware so running the flow isn't necessary.