SmartDesign: Added MSS wrapper SmartDesign
Added a wrapper SmartDesign around the MSS to clean up the canvas.

Added a command to mark memory map DRCs as warnings and not errors to
prevent a generation failue due to the mixed address space of FICs.

Removed the "FIC3_INITIATOR.tcl" script as this is no longer used.