Tcl: Use wildcards for SgCores
Updated the design Tcl scripts to use wildcards for SgCores as these are bundled with Libero and can be version dependant.

This will use the latest version of the core that is available and the latest version will be downloaded if not present in the local vault.

This may cause an issue if core parameters are added but this is not expected for these cores - if a parameter changes the core configuration Tcl script will have to be updated.