Fixed Windows issue with paths finding shls command
The user can now specify the path to a particular installation of SmartHLS using a script_args parameter.
This is particularly useful for incremental releases. If none provided,
then the script will attempt to get the path to shls from the PATH environment
variable, and if that fails it will try to find the version installed
with Libero.

Also fixed minor typo in readme.md