AXI4 Stream Demo: bugfix + datawith change
Change the datawidth used for the CoreAXI4DMAController in
the AXI4Stream demo example to be 64'b from 32'b. This change allows the
CoreAXI4DMAController to transfer data at a higher maximum rate.

Resolve a bug with the AXI Stream data generator module which was
incorrectly implementing the AXI4 Stream specification. See SAR #128874.

Corresponding updates to build script and BFM scripts as a result of
these changes.

Signed-off-by: Patrick Owens <Patrick.Owens@microchip.com>