Industrial Edge Demo: Adding two GPIOs for Motor Control functionality.
This demo requires to make use of Mikrobus connector for AN (direction)
and RST (Reset) pins to incorporate Stepper Motor functionality.
So added two new GPIOs, one for AN and another for RST.

In this demo, the UAExpert works as OPC UA Client and PolarFire SoC
ICICLE Kit as OPC UA Servers to communicate over OPC protocol. This
demo also explains the ICICLE capability to function as OPC UA Client
and OPC UA Server both.

Signed-off-by: Ahmed Mujtaba Ali <ahmed.ali2@microchip.com>